+++ R/geom-bar.R
@@ -61,7 +61,7 @@ GeomBarPattern <- ggproto(
   non_missing_aes = c("xmin", "xmax", "ymin", "ymax"),
 
   setup_params = function(data, params) {
-    params$flipped_aes <- has_flipped_aes(data, params, range_is_orthogonal = FALSE)
+    params$flipped_aes <- ggplot2::has_flipped_aes(data, params, range_is_orthogonal = FALSE)
     params
   },